+++ src/index.ts
@@ -14,6 +14,7 @@ export { default as ChipList } from './components/ChipList';
 export { default as TextFieldValidate } from './components/TextFieldValidate';
 export { default as TextFieldPassword } from './components/TextFieldPassword';
 export { default as TextFieldNumberValidate } from './components/TextFieldNumberValidate';
+export { default as TextFieldBirthDateWithAge } from './components/TextFieldBirthDateWithAge';
 
 export * from './utils/validateTelefone';
 export * from './utils/validateEmail';